refactor: improve performance and code flow
Improvements based on the code review: https://github.com/natura-cosmeticos/eslint-plugin-i18n-checker/pull/1#pullrequestreview-168535866
