+++ test/path-locator.spec.js
@@ -1,15 +1,16 @@
 const { assert } = require('chai');
 
-const pathLocalizator = require('../src/path-localizator');
+const PathLocator = require('../src/path-locator');
 
-describe('pathLocalizator', () => {
+describe('pathLocator', () => {
   describe('when path is a string (ex: "title")', () => {
     it('return undefined when path is found in locales files', () => {
       // given
       const locales = ['../test/fixtures/locales/es-pr.js'];
       const path = 'initialKits';
+      const pathLocator = new PathLocator(locales);
       // when
-      const result = pathLocalizator(path, locales);
+      const result = pathLocator.call(path);
 
       // then
       assert.isEmpty(result);
@@ -20,8 +21,9 @@ describe('pathLocalizator', () => {
       const filename = 'es-pr.js';
       const locales = [`../test/fixtures/locales/${filename}`];
       const path = 'notFoundString';
+      const pathLocator = new PathLocator(locales);
       // when
-      const result = pathLocalizator(path, locales);
+      const result = pathLocator.call(path);
 
       // then
       assert.isNotEmpty(result);
@@ -34,8 +36,9 @@ describe('pathLocalizator', () => {
       // given
       const locales = ['../test/fixtures/locales/es-pr.js'];
       const path = 'home.title';
+      const pathLocator = new PathLocator(locales);
       // when
-      const result = pathLocalizator(path, locales);
+      const result = pathLocator.call(path);
 
       // then
       assert.isEmpty(result);
@@ -46,8 +49,9 @@ describe('pathLocalizator', () => {
       const filename = 'es-pr.js';
       const locales = [`../test/fixtures/locales/${filename}`];
       const path = 'home.subtitle';
+      const pathLocator = new PathLocator(locales);
       // when
-      const result = pathLocalizator(path, locales);
+      const result = pathLocator.call(path);
 
       // then
       assert.isNotEmpty(result);
